requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pymongo==4.5.0
python-dotenv==1.0.0
pandas==2.1.0
numpy==1.25.2
scikit-learn==1.3.0
joblib==1.3.2
networkx==3.1
pydantic==2.3.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pytest==7.4.2
httpx==0.24.1
